requirements.txt
Django>=5.0
django-ninja>=1.1.0
PyJWT>=2.8.0
django-redis>=5.4.0
redis>=5.0.0
django-cors-headers>=4.3.0
